bug fixes + concurrent X11 threads
This commit updates the author.
It also contains the following changes:
- initializes Xlib support for concurrent threads
- fix minor device number calculation for large values
- fix crash if there are multiple pty slave folders

fix minor device number calculation for large values

Unfortunately the extraction of the minor device number from the tty_nr
was wrong.
The man page of procfs states that 'The minor device number is contained
[in the tty_nr] in the combination of bits 31 to 20 and 7 to 0' but it did not state
that you need to shift the bits together.
So I just missed that step.
This commit adds the correct placement of the bits to the calculation of the minor device number.

fix crash if there are multiple pty slave folders

If a system uses multiple pty slave folders it is possible that while
searching for the pty file a FileNotFoundError is raised which crashs
the program. In this case the correct behavior would be to look for the
pty file in the next folder. So this commit ignores the
FileNotFoundError for the call of the stat function.